removes redundant `<p>` tags
with the change to the get_all_posts() function to move the Markdown interpretation to where the $intro is defined, the home page adds a couple empty `<p></p>` tags, which results in the intro being indented via the basic css `p + p` rules. This removes the `<p>` tags from `/template/posts.php` since the `<p>` tags are being added by the Markdown anyway.