adds strikethrough support
this is a quick hack to support `~~` for strikethroughs, e.g. ~~strike~~. The hack part comes into play if your ~~ is at the end of a line, like: ~~this is a strikethrough.~~

In that case, you need to add an extra space at the end of the line so the `str_replace` function sees `~~ ` and replaces it with a `</del>`. If I get better at crazy preg_match or preg_split or other regular expression functions like that, this should be updated to support it more natively. but for now, this is probably more of an edge case just for me, , this should be update Mou supports ~~ as Strikethroughs.