Add bitly integration to dropplets
$short_url can be used in a theme in place of $post_link. It will
automatically convert the $post_link to a bitly link if the bitly
username and bitly api key are entered correctly in the settings.
Otherwise it falls back to $post_link. I have changed it in the "Share
on Twitter" and "Comment on Twitter" in the default theme as an example.