Cleaned Errors & Updated Readme
Took care of an issue occurring if the forward/back buttons were used while trying to view a private post. An undefined error was thrown for $_REQUEST['password'].

Took care of an issue with the write.php file, where it was not properly setting the value of fields when the post was New, only when an existing post was Edited.

Updated Readme to include quick setup instructions.